kafka/producer/requirements.txt
kafka-python-ng==2.2.3
zstandard==0.22.0
python-dotenv==1.0.1
